requirements.txt
altair
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
db-dtypes
gitdb
GitPython
google-api-core
google-auth
google-cloud-bigquery
google-cloud-core
google-crc32c==1.5.0
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
idna
Jinja2
joblib
jsonschema
jsonschema-specifications
markdown-it-py
MarkupSafe
mdurl
numpy
packaging
pandas
pillow
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pydeck
Pygments
python-dateutil
pytz
referencing
requests
rich
rpds-py
rsa
scikit-learn
scipy
six
smmap
splitfile
streamlit
tenacity
threadpoolctl
toml
toolz
tornado
typing_extensions
tzdata
urllib3
watchdog
gdown